+++ spec/features/early_career_payments_claim_spec.rb
@@ -58,11 +58,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     expect(claim.eligibility.reload.pgitt_or_ugitt_course).to eq "postgraduate"
 
-    # [PAGE 10] - ITT Subject
-    # - Which subject did you do your undergraduate ITT in
-    #   - OR -
-    # - Which subject did you do your postgraduate ITT in
-
+    # [PAGE 10] - Which subject did you do your postgraduate ITT in
     expect(page).to have_text(I18n.t("early_career_payments.questions.eligible_itt_subject", ug_or_pg: claim.eligibility.reload.pgitt_or_ugitt_course))
 
     choose "Mathematics"
@@ -70,6 +66,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     expect(claim.eligibility.reload.eligible_itt_subject).to eql "mathematics"
 
+    # [PAGE 11] - Which subject did you do your postgraduate ITT in
     # [PAGE 12] - Do you teach maths now
     expect(page).to have_text(I18n.t("early_career_payments.questions.teaching_subject_now", eligible_itt_subject: claim.eligibility.eligible_itt_subject))
 
@@ -78,11 +75,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     expect(claim.eligibility.reload.teaching_subject_now).to eql true
 
-    # [PAGE 13] - ITT Academic Year
-    # In what academic year did you start your undergraduate ITT
-    # - OR -
-    # In what academic year did you start your postgraduate ITT
-
+    # [PAGE 13] - In what academic year did you start your undergraduate ITT
     expect(page).to have_text(I18n.t("early_career_payments.questions.itt_academic_year", start_or_complete: "start", ug_or_pg: claim.eligibility.pgitt_or_ugitt_course))
 
     choose "2018 - 2019"
@@ -90,6 +83,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     expect(claim.eligibility.reload.itt_academic_year).to eql "2018_2019"
 
+    # [PAGE 14] - In what academic year did you start your postgraduate ITT
     # [PAGE 15] - Check your answers for eligibility
     expect(page).to have_text(I18n.t("early_career_payments.check_your_answers.part_one.primary_heading"))
     expect(page).to have_text(I18n.t("early_career_payments.check_your_answers.part_one.secondary_heading"))
@@ -106,8 +100,9 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     click_on("Continue")
 
-    # [PAGE 16] - You are eligible for an early-career payment
+    # [PAGE 16] - You are eligible for an early career payment
     expect(page).to have_text("You are eligible " + I18n.t("early_career_payments.claim_description"))
+
     within(".govuk-list--bullet") do
       expect(page).to have_text("your National Insurance number")
       ["bank account details", "teacher reference number", "QTS certificate", "student loan"].each do |bullet_point|
@@ -117,6 +112,9 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     click_on "Continue"
 
+    # [PAGE 20] - Personal Details
+    # [PAGE 21] - One Time Password
+    # [PAGE 22] - We have sent you reminders
     # [PAGE 23] - How will we use the information you provide
     expect(page).to have_text("How we will use the information you provide")
 
@@ -175,6 +173,14 @@ RSpec.feature "Teacher Early-Career Payments claims" do
     fill_in "claim_one_time_password", with: otp_in_mail_sent
 
     click_on "Confirm"
+
+    # Payment to Bank or Building Society
+    expect(page).to have_text(I18n.t("questions.bank_or_building_society"))
+
+    choose "Personal bank account"
+    click_on "Continue"
+
+    expect(claim.reload.bank_or_building_society).to eq "personal_bank_account"
 
     # [PAGE 27] - Enter bank account details
     expect(page).to have_text(I18n.t("questions.bank_details"))
@@ -253,7 +259,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
 
     expect(claim.eligibility.reload.postgraduate_doctoral_loan).to eql true
 
-    # TODO [PAGE 37] - Check your answers before sending your application
+    # [PAGE 37] - Check your answers before sending your application
     expect(page).to have_text("Check your answers before sending your application")
     expect(page).not_to have_text("Eligibility details")
     %w[Identity\ details Payment\ details Student\ loan\ details].each do |section_heading|
@@ -281,12 +287,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
     expect(page).to have_text(claim.reference)
   end
 
-  # As part of Eligible Later Process
-  # TODO [PAGE 20] - Personal Details
-  # TODO [PAGE 21] - One Time Password
-  # TODO [PAGE 22] - We have sent you reminders
-
-  scenario "Supply Teacher makes claim for 'Early-Career Payments' with a contract to teach for entire term & employed directly by school" do
+  scenario "Supply Teacher makes claim for 'Early Career Payments' with a contract to teach for entire term & employed directly by school" do
     visit landing_page_path(EarlyCareerPayments.routing_name)
     expect(page).to have_link(href: EarlyCareerPayments.feedback_url)