+++ spec/models/early_career_payments_spec.rb
@@ -26,7 +26,7 @@ RSpec.describe EarlyCareerPayments, type: :model do
 
   describe ".feedback_url" do
     it "returns a 'docs.google.com/forms/<slug>/viewform' url" do
-      # TODO get proper feedback URL - ECP-509
+      # get proper feedback URL - ECP-509
       expect(subject.feedback_url).to include("https://docs.google.com/forms/TO-BE-REPLACED-by-response-to-ECP-509/viewform")
     end
   end
@@ -48,7 +48,7 @@ RSpec.describe EarlyCareerPayments, type: :model do
       "3f85a1f7-9400-4b48-9a31-eaa643d6b977"
     end
     it "returns the notify_reply_to_id" do
-      # TODO replace with valid ID - ECP-515
+      # replace with valid ID - ECP-515
       expect(subject.notify_reply_to_id).to eql ecp_notify_reply_to_id
     end
   end